requirements.txt
aiogram==3.13.1
aiohttp==3.10.10
feedparser==6.0.11
python-dotenv==1.0.1
apscheduler==3.10.4
google-generativeai==0.8.3
openai==1.54.4
httpx==0.26.0